requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
